requirements.txt
requests
faiss-cpu
numpy
sentence-transformers
tqdm
